requirements.txt
langchain==0.2.10
langchain-community==0.2.10
langchain-chroma==0.1.4
langchain-huggingface==0.1.2

chromadb==0.5.0

llama-cpp-python==0.2.76

unstructured==0.15.9
pdfminer.six==20221105
pytesseract
pillow

torch
transformers
sentence-transformers

pdf2image
pdfplumber
camelot-py
tabula-py
opencv-python
numpy
tqdm

python-dotenv

